requirements.txt
docxtpl
jsonschema
boto3
